+++ include/astarte_device_sdk/device.h
@@ -19,6 +19,8 @@
  * @{
  */
 
+#include <zephyr/kernel.h>
+
 #include "astarte_device_sdk/astarte.h"
 #include "astarte_device_sdk/device_id.h"
 #include "astarte_device_sdk/individual.h"
@@ -180,11 +182,18 @@ astarte_result_t astarte_device_new(astarte_device_config_t *cfg, astarte_device
  * @brief Destroy the Astarte device instance.
  *
  * @note The device handle will become invalid after this operation.
+ * @note If the device is connected when calling this function it will be disconnected.
  *
  * @param[in] device Device instance to be destroyed.
+ * @param[in] timeout Timeout for this function when @p force is false.
+ * @param[in] force Force an eventual disconnection. When false this function will be blocking and
+ * wait untill all QoS 1 or larger messages have been successfully transmitted, or a timeout has
+ * been reached. When true this function will not wait for any message to be correctly delivered
+ * and force a device disconnection.
  * @return ASTARTE_RESULT_OK if successful, otherwise an error code.
  */
-astarte_result_t astarte_device_destroy(astarte_device_handle_t device);
+astarte_result_t astarte_device_destroy(
+    astarte_device_handle_t device, k_timeout_t timeout, bool force);
 
 /**
  * @brief add an interface to the device.
@@ -216,9 +225,15 @@ astarte_result_t astarte_device_connect(astarte_device_handle_t device);
  * @note It will be possible to re-connect the device after disconnection.
  *
  * @param[in] device Device instance to be disconnected.
+ * @param[in] timeout Timeout for this function when @p force is false.
+ * @param[in] force Force the disconnection. When false this function will be blocking and wait
+ * untill all QoS 1 or larger messages have been successfully transmitted, or a timeout has been
+ * reached. When true this function will not wait for any message to be correctly delivered and
+ * force a device disconnection.
  * @return ASTARTE_RESULT_OK if successful, otherwise an error code.
  */
-astarte_result_t astarte_device_disconnect(astarte_device_handle_t device);
+astarte_result_t astarte_device_disconnect(
+    astarte_device_handle_t device, k_timeout_t timeout, bool force);
 
 /**
  * @brief Poll data from Astarte.