+++ lib/astarte_device_sdk/device.c
@@ -189,11 +189,12 @@ failure:
     return ares;
 }
 
-astarte_result_t astarte_device_destroy(astarte_device_handle_t device)
+astarte_result_t astarte_device_destroy(
+    astarte_device_handle_t device, k_timeout_t timeout, bool force)
 {
     astarte_result_t ares = ASTARTE_RESULT_OK;
     if (device->connection_state != DEVICE_DISCONNECTED) {
-        ares = astarte_device_disconnect(device);
+        ares = astarte_device_disconnect(device, timeout, force);
         if (ares != ASTARTE_RESULT_OK) {
             return ares;
         }
@@ -223,9 +224,10 @@ astarte_result_t astarte_device_connect(astarte_device_handle_t device)
     return astarte_device_connection_connect(device);
 }
 
-astarte_result_t astarte_device_disconnect(astarte_device_handle_t device)
+astarte_result_t astarte_device_disconnect(
+    astarte_device_handle_t device, k_timeout_t timeout, bool force)
 {
-    return astarte_device_connection_disconnect(device);
+    return astarte_device_connection_disconnect(device, timeout, force);
 }
 
 astarte_result_t astarte_device_poll(astarte_device_handle_t device)