+++ CHANGELOG.md
@@ -10,6 +10,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - Unreleased
+### Changed
+- The `astarte_device_disconnect` function is not blocking when force is set to false.
+  It blocks until all outgoing/incoming messages are processed successfully or the timeout is
+  reached.
+
 ## [0.7.2] - 2024-10-23
 ### Changed
 - All the library samples have been merged into a single sample called `astarte_app`.